Fix type casts warnings. Add more logs to CMakeLists.txt

+++ include/entry.h
@@ -22,7 +22,7 @@ namespace btree {
 
         const K key;
         const ValueType data;
-        const int size_in_bytes;
+        const int32_t size_in_bytes;
 
         template <typename U = V, enable_if_t<std::is_arithmetic_v<U>> = true>
         explicit Entry() :
@@ -46,19 +46,19 @@ namespace btree {
         Entry(const K& key, const V value) :
                 key(key),
                 data(value),
-                size_in_bytes(sizeof(value)) {}
+                size_in_bytes(static_cast<int32_t>(sizeof(value))) {}
 
         template <typename U = V, enable_if_t<is_string_v<U>> = true>
         Entry(const K& key, const V& value) :
                 key(key),
                 data(reinterpret_cast<const ValueType>(value.c_str())),
-                size_in_bytes(value.size() * sizeof(typename V::value_type)) {}
+                size_in_bytes(static_cast<int32_t>(value.size() * sizeof(typename V::value_type))) {}
 
         template <typename U = V, enable_if_t<std::is_pointer_v<U>> = true>
         Entry(const K& key, const V& data, const int size) :
                 key(key),
                 data(reinterpret_cast<const ValueType>(data)),
-                size_in_bytes(size * size_of_original_value_type)
+                size_in_bytes(static_cast<int32_t>(size * size_of_original_value_type))
         {
             typedef typename std::remove_pointer_t<V> data_type;
             static_assert(sizeof(data_type) == sizeof(OriginalValueType));
@@ -106,8 +106,8 @@ namespace btree {
             }
         }
 #ifdef UNIT_TESTS
-        int size_in_file() {
-            int size = sizeof(K);
+        int32_t size_in_file() {
+            int32_t size = sizeof(K);
             if constexpr (is_string_v<V> || V_is_pointer) {
                 size += 4; // strlen
             }
